+++ src/test/fetchRockets.test.js
@@ -1,8 +1,6 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import axios from 'axios';
 import { fetchRockets } from '../redux/rocketsSlice/rocketsSlice';
-
 
 // Create a mock store
 const mockStore = configureMockStore([thunk]);
@@ -19,8 +17,12 @@ describe('FetchRockets async action', () => {
     // Mock the fetch function to return a successful response
     global.fetch = jest.fn().mockResolvedValue({
       json: () => Promise.resolve([
-        { id: '1', name: 'Falcon 1', description: 'Description 1',type:'none', flickrimages:'',reserved:false},
-        { id: '1', name: 'Falcon 9', description: 'Description 1',type:'none', flickrimages:'',reserved:false},
+        {
+          id: '1', name: 'Falcon 1', description: 'Description 1', type: 'none', flickrimages: '', reserved: false,
+        },
+        {
+          id: '1', name: 'Falcon 9', description: 'Description 1', type: 'none', flickrimages: '', reserved: false,
+        },
       ]),
     });
 
@@ -28,11 +30,15 @@ describe('FetchRockets async action', () => {
     const expectedActions = [
       fetchRockets.pending().type,
       fetchRockets.fulfilled([
-        { id: '1', name: 'Falcon 1', description: 'Description 1',type:'none', flickrimages:'',reserved:false},
-        { id: '1', name: 'Falcon 9', description: 'Description 1',type:'none', flickrimages:'',reserved:false},
+        {
+          id: '1', name: 'Falcon 1', description: 'Description 1', type: 'none', flickrimages: '', reserved: false,
+        },
+        {
+          id: '1', name: 'Falcon 9', description: 'Description 1', type: 'none', flickrimages: '', reserved: false,
+        },
       ]).type,
     ];
-   
+
     // Dispatch the async action
     await store.dispatch(fetchRockets());
 
@@ -43,7 +49,7 @@ describe('FetchRockets async action', () => {
     expect(dispatchedActions.map((action) => action.type)).toEqual(expectedActions);
   });
 
- /*  it('dispatches the correct actions when API call fails', async () => {
+  /*  it('dispatches the correct actions when API call fails', async () => {
     // Mock the fetch function to throw an error
     global.fetch = jest.fn().mockRejectedValue('API error');
